Implemented discounts chaining with parent propepry
In case formula is multiline (contains more than one discount),
it was impossible to have multiple discounts attached to a single price
beacuse of uniqness constraints.

Considering every next formula is applied to the previous formula
calculation results, it seems wise to chain discounts, so parent_id
changes and uniq constarint is satisfied.