When leftCharge gets replaced, the rightCharges parent should be reassigned only if it existed before
Originally fixes the problem when there are two formulas:

discount.fixed('100%').till('06.2020')
leasing.since('08.2019').lasts('12 months')

When evaluated in 06.2020, the first formula is ignored
The second formula replaces the original monthly,hardware with monthly,leasing charge and
it should not have a parent. Instead, parent was forcefulle created, leading to charge
duplication based on UniqueId.