Fixes stuck clock in a transaction
Calculator can be created at the beginning of a transaction (1)
in order to calculate it's expected amount.

Then the transaction (sale, bill, charges) can be persisted (2)
and the Calculator CAN BE reused for new calculation.

In this case, the newer calculations (3) will happen with a stuck at
point ^1.

After a commit, the actual clock is used by default. CachedDateTimeProvider is provided for Unit testing.